refactor: remove __init__ in file_interface
