--- oxspires_tools/trajectory/file_interfaces/__init__.py
@@ -1,5 +1,0 @@
-# ruff: noqa: F401
-from .nerf import NeRFTrajReader, NeRFTrajWriter
-from .timestamp import TimeStamp
-from .tum import TUMTrajReader, TUMTrajWriter
-from .vilens_slam import VilensSlamTrajReader
