+++ tests/test_timestamp.py
@@ -1,7 +1,7 @@
 import numpy as np
 import pytest
 
-from oxspires_tools.trajectory.file_interfaces import TimeStamp
+from oxspires_tools.trajectory.file_interfaces.timestamp import TimeStamp
 
 
 def test_init_with_sec_nsec():
